* added a global lua-scope across all connections * learnt about lua_newthread(), luaL_ref() and _unref()
instead of closing the lua_State after each connection we create a 
new lua-thread which gets its global scope from the the global
lua_State and has its own exec-stack and script-env.

This has the nice side-effect that we only use one lua-state which is 
initialized once. The proxy.* namespace is global and shared between
all instances. Only proxy.response is per connection.