reworked the connection pool to have one pool for user (fixes #....)
before we had one connection pool which stored idling connections for
all users. In disconnect_client() we tried to shrink the pool to
max-idle and could specify for which user we want to close connections

Now this mgmt is move into the core and only the configuration is
external. To reflect this internal change the scripting-API changes too:

  proxy.backend[ndx].
    idling_connections 

is replaced by

  proxy.backend[ndx].
    pool.
      min_idle_connections (rw)
      max_idle_connections (rw)
      users[username].
        cur_idle_connections (ro)

Another important change handles the order of getting connections from
the pool. Before we picked one randomly, now we pick the oldest in the
pool to reduce the impact of wait_timeout killing old, idling
connections